feat(web): move agent webapp access control entry into action row button
