+++ web/features/agent-v2/agent-detail/access/components/web-app-access-card.tsx
@@ -17,7 +17,7 @@ import { AccessMode } from '@/models/access-control'
 import { consoleQuery } from '@/service/client'
 import { AppModeEnum } from '@/types/app'
 import { accessSurfaceActionClassName, AccessSurfaceCard } from './access-surface-card'
-import { WebAppAccessControlSection } from './web-app-access-control-section'
+import { WebAppAccessControlButton } from './web-app-access-control-button'
 
 export function WebAppAccessCard({
   agent,
@@ -220,10 +220,6 @@ export function WebAppAccessCard({
       disabled={isLoading || !canManageWebApp}
       busy={isBusy}
     >
-      <WebAppAccessControlSection
-        agent={agent}
-        onAccessModeUpdated={() => queryClient.invalidateQueries({ queryKey: agentDetailQueryKey })}
-      />
       {webAppUrl && isEnabled ? (
         <a
           href={webAppUrl}
@@ -271,6 +267,10 @@ export function WebAppAccessCard({
         <span aria-hidden className="i-ri-palette-line size-4" />
         {t(($) => $['agentDetail.access.webApp.actions.settings'])}
       </Button>
+      <WebAppAccessControlButton
+        agent={agent}
+        onAccessModeUpdated={() => queryClient.invalidateQueries({ queryKey: agentDetailQueryKey })}
+      />
       {settingsAppInfo && (
         <SettingsModal
           isChat
