+++ web/features/agent-v2/agent-detail/access/components/web-app-access-control-button.tsx
@@ -1,0 +1,64 @@
+'use client'
+
+import type { AgentAppDetailWithSite } from '@dify/contracts/api/console/agent/types.gen'
+import { Button } from '@langgenius/dify-ui/button'
+import { useQuery } from '@tanstack/react-query'
+import { useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import AccessControl from '@/app/components/app/app-access-control'
+import { systemFeaturesQueryOptions } from '@/features/system-features/client'
+import { useAppACLCapabilities } from '@/hooks/use-app-acl-capabilities'
+import { AccessMode, isAccessMode } from '@/models/access-control'
+
+export function WebAppAccessControlButton({
+  agent,
+  onAccessModeUpdated,
+}: {
+  agent?: AgentAppDetailWithSite
+  onAccessModeUpdated: () => void | Promise<void>
+}) {
+  const { t } = useTranslation('agentV2')
+  const [showAccessControl, setShowAccessControl] = useState(false)
+  const appId = agent?.app_id
+  const rawAccessMode = agent?.access_mode
+  const accessMode = isAccessMode(rawAccessMode)
+    ? rawAccessMode
+    : AccessMode.SPECIFIC_GROUPS_MEMBERS
+  const { data: systemFeatures } = useQuery(systemFeaturesQueryOptions())
+  const { canReleaseAndVersion: canManageWebAppAccessControl } = useAppACLCapabilities(
+    agent?.permission_keys,
+    agent?.maintainer,
+  )
+
+  if (!systemFeatures?.webapp_auth.enabled || !canManageWebAppAccessControl || !appId) return null
+
+  const handleConfirm = async () => {
+    setShowAccessControl(false)
+    try {
+      await onAccessModeUpdated()
+    } catch (error) {
+      console.error('Failed to refresh agent detail:', error)
+    }
+  }
+
+  return (
+    <>
+      <Button
+        variant="secondary"
+        size="medium"
+        className="gap-1.5 px-3"
+        onClick={() => setShowAccessControl(true)}
+      >
+        <span aria-hidden className="i-ri-lock-2-line size-4" />
+        {t(($) => $['agentDetail.access.webApp.actions.accessControl'])}
+      </Button>
+      {showAccessControl && (
+        <AccessControl
+          app={{ id: appId, access_mode: accessMode }}
+          onClose={() => setShowAccessControl(false)}
+          onConfirm={handleConfirm}
+        />
+      )}
+    </>
+  )
+}
