+++ web/i18n/ar-TN/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "نقطة الوصول",
   "agentDetail.access.toggleSurface": "تبديل وصول {{name}}",
   "agentDetail.access.webApp.accessUrl": "رابط الوصول",
+  "agentDetail.access.webApp.actions.accessControl": "التحكم في الوصول",
   "agentDetail.access.webApp.actions.customize": "واجهة أمامية مخصصة",
   "agentDetail.access.webApp.actions.embedded": "مضمّن",
   "agentDetail.access.webApp.actions.launch": "تشغيل",
