+++ web/i18n/es-ES/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Punto de acceso",
   "agentDetail.access.toggleSurface": "Alternar acceso de {{name}}",
   "agentDetail.access.webApp.accessUrl": "URL de acceso",
+  "agentDetail.access.webApp.actions.accessControl": "Control de Acceso",
   "agentDetail.access.webApp.actions.customize": "Frontend personalizado",
   "agentDetail.access.webApp.actions.embedded": "Incrustado",
   "agentDetail.access.webApp.actions.launch": "Iniciar",
