+++ web/i18n/id-ID/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Titik Akses",
   "agentDetail.access.toggleSurface": "Alihkan akses {{name}}",
   "agentDetail.access.webApp.accessUrl": "URL Akses",
+  "agentDetail.access.webApp.actions.accessControl": "Kontrol Akses",
   "agentDetail.access.webApp.actions.customize": "Frontend Kustom",
   "agentDetail.access.webApp.actions.embedded": "Tertanam",
   "agentDetail.access.webApp.actions.launch": "Luncurkan",
