+++ web/i18n/nl-NL/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Toegangspunt",
   "agentDetail.access.toggleSurface": "Toegang van {{name}} schakelen",
   "agentDetail.access.webApp.accessUrl": "Toegangs-URL",
+  "agentDetail.access.webApp.actions.accessControl": "Toegangsbeheer",
   "agentDetail.access.webApp.actions.customize": "Aangepaste frontend",
   "agentDetail.access.webApp.actions.embedded": "Ingesloten",
   "agentDetail.access.webApp.actions.launch": "Starten",
