+++ web/i18n/pl-PL/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Punkt dostępu",
   "agentDetail.access.toggleSurface": "Przełącz dostęp {{name}}",
   "agentDetail.access.webApp.accessUrl": "URL dostępu",
+  "agentDetail.access.webApp.actions.accessControl": "Kontrola dostępu",
   "agentDetail.access.webApp.actions.customize": "Niestandardowy frontend",
   "agentDetail.access.webApp.actions.embedded": "Osadzony",
   "agentDetail.access.webApp.actions.launch": "Uruchom",
