+++ web/i18n/ru-RU/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Точка доступа",
   "agentDetail.access.toggleSurface": "Переключить доступ {{name}}",
   "agentDetail.access.webApp.accessUrl": "URL доступа",
+  "agentDetail.access.webApp.actions.accessControl": "Управление доступом",
   "agentDetail.access.webApp.actions.customize": "Пользовательский фронтенд",
   "agentDetail.access.webApp.actions.embedded": "Встроить",
   "agentDetail.access.webApp.actions.launch": "Запустить",
