+++ web/i18n/vi-VN/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "Điểm truy cập",
   "agentDetail.access.toggleSurface": "Bật/tắt truy cập {{name}}",
   "agentDetail.access.webApp.accessUrl": "URL truy cập",
+  "agentDetail.access.webApp.actions.accessControl": "Kiểm soát truy cập",
   "agentDetail.access.webApp.actions.customize": "Giao diện tùy chỉnh",
   "agentDetail.access.webApp.actions.embedded": "Nhúng",
   "agentDetail.access.webApp.actions.launch": "Khởi chạy",
