+++ web/i18n/zh-Hans/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "访问点",
   "agentDetail.access.toggleSurface": "切换 {{name}} 访问状态",
   "agentDetail.access.webApp.accessUrl": "访问 URL",
+  "agentDetail.access.webApp.actions.accessControl": "访问控制",
   "agentDetail.access.webApp.actions.customize": "自定义前端",
   "agentDetail.access.webApp.actions.embedded": "嵌入",
   "agentDetail.access.webApp.actions.launch": "启动",
