+++ web/i18n/zh-Hant/agent-v-2.json
@@ -14,6 +14,7 @@
   "agentDetail.access.title": "存取點",
   "agentDetail.access.toggleSurface": "切換 {{name}} 存取狀態",
   "agentDetail.access.webApp.accessUrl": "存取 URL",
+  "agentDetail.access.webApp.actions.accessControl": "存取控制",
   "agentDetail.access.webApp.actions.customize": "自訂前端",
   "agentDetail.access.webApp.actions.embedded": "嵌入",
   "agentDetail.access.webApp.actions.launch": "啟動",
